requirements.dev.txt
black==25.1.0
flake8==7.2.0
isort==5.13.2
pre-commit==4.2.0
pytest==8.4.0
pytest-django==4.11.1
